capstone::arch

Enum ArchDetail

pub enum ArchDetail<'a> {

    ArmDetail(ArmInsnDetail<'a>),
    Arm64Detail(Arm64InsnDetail<'a>),
    EvmDetail(EvmInsnDetail<'a>),
    M680xDetail(M680xInsnDetail<'a>),
    M68kDetail(M68kInsnDetail<'a>),
    MipsDetail(MipsInsnDetail<'a>),
    Mos65xxDetail(Mos65xxInsnDetail<'a>),
    PpcDetail(PpcInsnDetail<'a>),
    RiscVDetail(RiscVInsnDetail<'a>),
    ShDetail(ShInsnDetail<'a>),
    SparcDetail(SparcInsnDetail<'a>),
    Tms320c64xDetail(Tms320c64xInsnDetail<'a>),
    TriCoreDetail(TriCoreInsnDetail<'a>),
    X86Detail(X86InsnDetail<'a>),
    XcoreDetail(XcoreInsnDetail<'a>),
    BpfDetail(BpfInsnDetail<'a>),
    SysZDetail(SysZInsnDetail<'a>),

}

Contains architecture-dependent detail structures.

For convenience, there are methods for each architecture that return an Option of that architecture’s detail structure. This allows you to use an if let Some(...) = { /* ... */ } instead of a match statement.

Variants

ArmDetail(ArmInsnDetail<'a>)

Arm64Detail(Arm64InsnDetail<'a>)

EvmDetail(EvmInsnDetail<'a>)

M680xDetail(M680xInsnDetail<'a>)

M68kDetail(M68kInsnDetail<'a>)

MipsDetail(MipsInsnDetail<'a>)

Mos65xxDetail(Mos65xxInsnDetail<'a>)

PpcDetail(PpcInsnDetail<'a>)

RiscVDetail(RiscVInsnDetail<'a>)

ShDetail(ShInsnDetail<'a>)

SparcDetail(SparcInsnDetail<'a>)

Tms320c64xDetail(Tms320c64xInsnDetail<'a>)

TriCoreDetail(TriCoreInsnDetail<'a>)

X86Detail(X86InsnDetail<'a>)

XcoreDetail(XcoreInsnDetail<'a>)

BpfDetail(BpfInsnDetail<'a>)

SysZDetail(SysZInsnDetail<'a>)

Implementations

impl<'a> ArchDetail<'a>

pub fn operands(&'a self) -> Vec<ArchOperand>

Returns architecture independent set of operands

pub fn arm(&'a self) -> Option<&'a ArmInsnDetail<'a>>

Returns the ARM details, if any

pub fn arm64(&'a self) -> Option<&'a Arm64InsnDetail<'a>>

Returns the ARM64 details, if any

pub fn evm(&'a self) -> Option<&'a EvmInsnDetail<'a>>

Returns the EVM details, if any

pub fn m680x(&'a self) -> Option<&'a M680xInsnDetail<'a>>

Returns the M680X details, if any

pub fn m68k(&'a self) -> Option<&'a M68kInsnDetail<'a>>

Returns the M68K details, if any

pub fn mips(&'a self) -> Option<&'a MipsInsnDetail<'a>>

Returns the MIPS details, if any

pub fn mos65xx(&'a self) -> Option<&'a Mos65xxInsnDetail<'a>>

Returns the Mos65xx details, if any

pub fn ppc(&'a self) -> Option<&'a PpcInsnDetail<'a>>

Returns the PPC details, if any

pub fn riscv(&'a self) -> Option<&'a RiscVInsnDetail<'a>>

Returns the RISCV details, if any

pub fn sh(&'a self) -> Option<&'a ShInsnDetail<'a>>

Returns the SH details, if any

pub fn sparc(&'a self) -> Option<&'a SparcInsnDetail<'a>>

Returns the SPARC details, if any

pub fn tms320c64x(&'a self) -> Option<&'a Tms320c64xInsnDetail<'a>>

Returns the Tms320c64x details, if any

pub fn tricore(&'a self) -> Option<&'a TriCoreInsnDetail<'a>>

Returns the TriCore details, if any

pub fn x86(&'a self) -> Option<&'a X86InsnDetail<'a>>

Returns the X86 details, if any

pub fn xcore(&'a self) -> Option<&'a XcoreInsnDetail<'a>>

Returns the XCore details, if any

pub fn bpf(&'a self) -> Option<&'a BpfInsnDetail<'a>>

Returns the BPF details, if any

pub fn sysz(&'a self) -> Option<&'a SysZInsnDetail<'a>>

Returns the SysZ details, if any